requirements.txt
flask
flask_cors
gunicorn
openai>=1.0.0
pillow